chore: unify component for job creation
